==> crawl_naver_news_title/3_analyze.py <==
""" 데이터 분석 """
import pandas as pd
import os
import mecab
import numpy as np
from collections import Counter
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)
import time
import joblib

base = os.path.dirname(os.path.abspath(__file__)) # 경로설정
df = pd.read_csv(os.path.join(base,'data','news_crawl_all.csv'))

# 한글과 공백을 제외하고 모두 제거한 특성 생성
df['text'] = df['title'].str.replace("[^ㄱ-ㅎㅏ-ㅣ가-힣 ]","")
print("중복 제거 전 shape:",df.shape)
# 제목 중복 제거 (트렌드를 파악하는데 도움이 되는가? -> 단순히 기사가 많이 나온것은 트렌드가 아니라고 판단함)
df.drop_duplicates(subset = ['text'], inplace=True)
print("중복 제거 후 shape:",df.shape)
# 결측치 확인
#df['text'].replace('', np.nan, inplace=True)
#print(df.isnull().sum())
#print(df)
mecab = mecab.MeCab()
# 식품 산업의 트렌드에 적합하지 않은 단어 제거 (사업, 정부, 지역, 업체, 부사, 조사 등) (해외는 제외)
# 선정한 방법 : 전체 데이터셋 토큰화한 단어들 중 상위 100 빈도를 가진 단어들을 확인하며 직접 제거 (트렌드와 연관성이 아예 없는것들)
stopwords = ['에','공사','한국','는','업계','협약','도','수출','과','산업','의','로','한','화','은','기업','출시','년','성','사업','부','주','와'\
            ,'식','약','가','등','업무','센터','체결','업체','강화','하','관리','이','선정','인증','서울','형','서','장관','억','다','처','원'\
            ,'기관','고','만','시장','모집','신','실천','기술','점검','제품','식품','유통','농수산','기능','안전','푸드','농림','가공','축산'\
            ,'물가','제조','수산','개발','지원','우수','공정','공장','가격','개최','판매','세계','일','식생활','까지','아','대','시','으로'\
            ,'월','위','광고','흥원','대상','확대','진','클러스터','중단','추진','농산물','가정','경영','기부','진행','인','주가','전','비','생산'\
            ,'건','브랜드','첫','주년','캠페인','맞','달','상승','교육청','투자','전남','인니','나','협력','스폰서','콘서트','경북','모색','채용'\
            ,'특징','활용','전문','참여','해','농','배','자','맛','국제','준공','평가','인턴','집중','최고','미래','종합','전북','부터','어','사용'\
            ,'할인','울산','대구','국내','교육','강세','국가','공모','박차','박람회','포토','내','스마트','확보','실시','명','획득','선','진단'\
            ,'오픈','한다','메인','위한','개','트렌드','업소','먹거리','육성','새','달성','플랜트','부당','위해','안','방안','빌','설성','행사'\
            ,'금지','먹','코로나','식품위생','기','맞춤','공략','서울시','창업','콘텐츠','경쟁력','후보자','경기도','최초','론','걸음','차세대'\
            ,'입','최강전','적발','칭','분야','에서','가치','글로벌','연계','계','더','및','간','정보','최대','옥천군','적','연구','데','맞손'\
            ,'검사','을','협회','학교','곳','확산','후보','오늘','사','소비자','제회','우리','케어','기호','씨','들','품','판로','플랫','올해'\
            ,'경남','폼','없','조성','날','제','미국','달러','물류','품질','신고','종','인기','삼성전자','받','정부','구매','증가','허가','용'\
            ,'전달','지','매출','현장','잡','선물','가능','개선','제로','세대','단','대표','게','밀','를','단지','농업','제주','할','밀양'\
            ,'소식','정책','성장','급등','경','불법행위','인상','사회','외식업','불','피해','백신','뉴스','운영','신청','실적','경제','전망'\
            ,'동서','삼양','불량','웅진','위생','위반','추석','회장','삼립','좋','롯데','대전','위생법','성수','농협','관','온라인'\
            ,'지역','수상','쇼핑','천','참가','특별','제일제당','상','표시','의약품','중','맥심','홈','단속','진출','보','있','풀무원'\
            ,'설','인사','호','세요','오','가지','부산','나눔','원료','장','정','명절','표','산','이벤트','샘','성분','백화점','연구원'\
            ,'세트','식약청','과학','익산','품안','대한민국','점','음식','외식','시설','대비','엑스포','충남','마케팅','검출','효과','전처'\
            ,'예방','생','문화','홍보','마트','사장','회사','상품','대회','마켓','차','알','음료','기준','재','인천','공개','소비'\
            ,'기한','방문','니','되','주변','는다','전국','계약','몰','이마트','현대','관련','수','영업','찾','그룹','봬','충북','된','엔'\
            ,'것','사랑','카누','활동','속','신세계','함께','회수','회','연구소','조사','허위','지도','열','물질','규모','대형','활성'\
            ,'동원','접객업소','전년','지정','해야','생활','교수','국민','노인','여름','삼','본격','기획','소재','라','부문','인수','불법'\
            ,'조','서비스','논란','화장품','급식','주목','법','앤','합동','농촌','백','체험','두','신규','편식','홈플러스','경찰','섭취'\
            ,'롯데마트','시스템','제약','앞두','집','내년','처장','네','도입','규제','혁신','근절','선보여','부정','광주','분석','모델'\
            ,'여름철','권','었','눈','경기','신송','돌파','자연','기탁','농심','용품','구축','여성','인천시','반','발표','지난해','확인','발전'\
            ,'열품','명회','보다','원산지','만드','중소','개막','효능','요리','유발','열풍','전략','검거','제공','얼','추가','의무','리뉴'\
            ,'돕','보조','알레르기','시대','공급','연속','팔','스','복지','에게','학회','유해','세','다양','고객','접수','오뚜기','맞이','잘'\
            ,'아이','오리온','급증','나선다','무','류','해태','미','시행','올','융자','생명','과정','이유','아라','늘','눈길','구속','주식'\
            ,'일당','보건','일자리','감시원','기대','동물','식품명인','업자','포럼','공동','만든','으면','허용','내달','조치','의원','일제'\
            ,'간담회','표창','폭행','종목','무더기','리얼','포스트','판','값','천안','이상','담','도움','단신','통합','소','만들','봉사','여'\
            ,'식중독','도약','그린','넣','높','많','운전기사','본부','선임','대장균','기념','분기','학진','안전성','진흥','개소','발간'\
            ,'치료제','천만','혐의','공업','이사','길','물','리','김','용기','행정','유치','이동필','넘','작년','계층','위원회','인정','이용'\
            ,'가짜','명인','상담','특화','자원','경쟁','후원','함유','입맛','개척','초','관심','상반기','별','일반','성과','직구','입주'\
            ,'음식점','계획','코리아','대학','확진','마련','매일','최우수','부산시','육','전쟁','영','족','전시회','가장','프로','째','감소'\
            ,'결정','양성','의료','특사','결과','우려','질','횡령','티','맛있','저','티몬','전주국','살','시작','취급','지수','큰','대책'\
            ,'자금','수준','않','전문가','검찰','률','한성','대통령','한류','과장','명품','강원','주요','발','타','영양학','못','돈','광주시'\
            ,'전주','스타','피부','천호','롯데제과','발탁','청년','전환','지속','뜬다','쿠팡','김재수','이웃','담회','남성','원장','후','그'\
            ,'겟','정기','예산','이색','번','겠','필요','이개호','성공','신설','증권','상당','설립','입건','아침','메','상대','방','말'\
            ,'목표','주의','된다','협의회','품목','재료','직원','개월','통해','안정','손잡','차단','과대광고','취약','프','뷰티','왜','외'\
            ,'단독','자리','세미나','가구','속여','김만식','현재','익','베지','손','슈퍼','청','또','투','방역','골드','적용','제도'\
            ,'바람','힘','명예','기금','업종','마감','쇼핑몰','데이','키','웰','수사','하나','대세','사범','등록','중소기업','공학','연말'\
            ,'거점','무료','양','사고','공정위','인터넷','약국','돼','최','감사','군','하락','시험','환자','비상','방송','어려','휴','축산물'\
            ,'프레시','때','수산물','전용','등급','입점','치료','정보원','협업','변신','식물','던','친화','식탁','지난','패션','중앙','기소'\
            ,'트럭','직매장','스킨','뱅크','힐링','올가','홀','페스티벌','컬','개장','플랜','축제','테크','힐링','매장','완주','메뉴','백종원'\
            ,'장터','쇼','행복','파이터','코트','셰프','와일드','트립','해마','레시피','직거래','파스퇴르','식자재','농가','완주군','패스트푸드'\
            ,'나무','맘','희망','세종','페어','피','소울','순천','나잇','스타트업','종시','웨이','바','머스','즐기','추천','합병','리포트'\
            ,'터','푸','대박','쓰','코엑스','아트','케이','라인','킹','아워홈','도전','테라','수립','문','스타일리스트','전문점','죠','몸'\
            ,'밥상','마을','공원','카드','벅스','매수','이야기','스토리','맥도날드','온','아모제','뜨','의성','식당','프로그램','남','하림'\
            ,'시민','가족','준비','사진','남양주','기자','만나','협동조합']

print('토큰화 진행 중 . . .')
start = time.time()
df['tokenized'] = df['text'].apply(mecab.morphs)
df['tokenized'] = df['tokenized'].apply(lambda x: [item for item in x if item not in stopwords])

# 빈도검증(stopwords 생성에 활용) 
# 토큰화 후 빈도 높은 단어 검증
#words = np.hstack(df['tokenized'].values)
#word_count = Counter(words)
#print(word_count.most_common(100))

# 데이터프레임 정렬
df = df.reset_index(drop=True)
df = df[['date','title','text','tokenized']]

# 토큰화한 데이터프레임 csv로 저장
#df.to_csv(os.path.join(base,'data','news_crawl_all_tok.csv'))
# CSV로 넘기면 tokenized 내부가 리스트가 아닌 문자열이 되서 오류발생, Pickle화 한 데이터프레임을 고대로 넘겨서 해결

# 데이터프레임 pickle화
joblib.dump(df, os.path.join(base,'data','df'))

# 소요시간
print("토큰화 완료 \nTimeSpent :", round(time.time() - start,2),"sec")
